+++ components/NoDataPlaceholder.tsx
@@ -1,0 +1,7 @@
+export default function NoDataPlaceHolder() {
+  return (
+    <div className="flex flex-row min-h-screen justify-center items-center">
+      No Blogs to show !!
+    </div>
+  );
+}
