+++ components/PageErrorState.tsx
@@ -1,0 +1,7 @@
+export default function PageErrorState() {
+  return (
+    <div className="flex flex-row min-h-screen justify-center items-center">
+      Ohh NO!! Something went wrong
+    </div>
+  );
+}
